Added docs for BaseBottomSheet.kt, ActionBottomSheet.kt and ActionBottomSheetButton.kt
