+++ app/src/main/java/com/design/composeChilli/MainActivity.kt
@@ -7,27 +7,19 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.design.composeChilli.ui.theme.NurComposeChiliTheme
-import com.design.composechili.components.bottom_sheet.ActionBottomSheet
-import com.design.composechili.components.bottom_sheet.ActionBottomSheetButton
-import com.design.composechili.components.bottom_sheet.ActionBottomSheetButtonType
-import com.design.composechili.components.bottom_sheet.ActionBottomSheetParams
+import com.design.composechili.components.bottomSheet.actionBottomSheet.ActionBottomSheet
+import com.design.composechili.components.bottomSheet.actionBottomSheet.ActionBottomSheetParams
 import com.design.composechili.components.cell.BaseCell
-import com.design.composechili.components.input.inputFieldWithDescAndAction.InputFieldWithDescAndAction
-import com.design.composechili.components.tooltip.ChiliTooltip
 import com.design.composechili.extensions.getBottomSheetState
 import com.design.composechili.theme.ChiliTheme
 import kotlinx.coroutines.launch
