+++ composeChili/src/main/java/com/design/composechili/components/bottomSheet/actionBottomSheet/ActionBottomSheetButton.kt
@@ -1,13 +1,22 @@
-package com.design.composechili.components.bottom_sheet
+package com.design.composechili.components.bottomSheet.actionBottomSheet
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.tooling.preview.Preview
+import com.design.composechili.R
 import com.design.composechili.components.buttons.baseButton.BaseButton
 import com.design.composechili.components.buttons.baseButton.ChiliButtonStyle
 import com.design.composechili.theme.ChiliTextStyle
 import com.design.composechili.theme.ChiliTheme
+
+/**
+ * Used for [ActionBottomSheet], displayed on BottomSheet with LazyColumn
+ * @param [actionBottomSheetParams] ActionBottomSheetButton visual transformation params and click listener
+ * @see [ActionBottomSheet]
+ */
 
 @Composable
 fun ActionBottomSheetButton(
@@ -15,7 +24,6 @@ fun ActionBottomSheetButton(
 ) {
 
     ChiliTheme {
-
         val titleStyle = actionBottomSheetParams.titleStyle ?: ChiliTextStyle.get(
             textSize = ChiliTheme.Attribute.ChiliTextDimensions.TextSizeH8,
             font = ChiliTheme.Attribute.ChiliBoldTextFont
@@ -25,7 +33,20 @@ fun ActionBottomSheetButton(
             onClick = { actionBottomSheetParams.onClick?.invoke() },
             title = actionBottomSheetParams.title,
             buttonStyle = ChiliButtonStyle.Secondary,
-            titleStyle = titleStyle.copy(actionBottomSheetParams.buttonTextColor)
+            titleStyle = titleStyle.copy(actionBottomSheetParams.buttonTextColor),
+        )
+    }
+}
+
+@Composable
+@Preview(showBackground = true)
+fun ActionBottomSheetButtonPreview() {
+    ChiliTheme {
+        ActionBottomSheetButton(
+            actionBottomSheetParams = ActionBottomSheetParams(
+                title = "TestTitle",
+                buttonTextColor = ChiliTheme.Colors.ChiliComponentButtonTextColorActive,
+            )
         )
     }
 }
