+++ composeChili/src/main/java/com/design/composechili/components/bottomSheet/baseBottomSheet/BaseBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.design.composechili.components.bottom_sheet
+package com.design.composechili.components.bottomSheet.baseBottomSheet
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedVisibility
@@ -20,18 +20,14 @@ import androidx.compose.material3.BottomSheetScaffold
 import androidx.compose.material3.BottomSheetScaffoldState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
@@ -40,6 +36,24 @@ import com.design.composechili.extensions.isExpanded
 import com.design.composechili.extensions.isExpanding
 import com.design.composechili.theme.ChiliTheme
 import kotlinx.coroutines.launch
+
+/**
+ * @param [modifier] Will be applied to sheetContent root composable
+ * @param [sheetState] Bottom sheet state, hosting state like expanded or not
+ ** @see [BottomSheetScaffoldState]
+ * @param [peekHeight] The default peek height used by BottomSheetScaffold.
+ * @param [hasCloseIcon] The icon which will show up on the top right corner of the sheet
+ * @param [collapseOnBackPressed] System back button handler enable/disable flag. Button will hide bottomSheet
+ * @param [baseBottomSheetParams] BottomSheet visual transformation params
+ * @param [isBackgroundDimmingEnabled] Background dimming enable/disable flag
+ * @param [bottomSheetSwipeEnabled] Bottom sheet swipe(drag) enable/disable flag
+ * @param [isDragHandleContentEnabled] Drag handle enable/disable flag
+ * @param [dragHandle] Drag handle composable content, by default is [BottomSheetDefaults.DragHandle].Will show on the top of the sheet
+ * @param [bottomSheetContent] BottomSheet content, which will show in the bottomSheet
+ * @param [screenContent] Screen content, which should be covered by the [BaseBottomSheet]. Bottom Sheet will show over this content
+ * @sample [BaseBottomSheetParams]
+ */
+
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -71,7 +85,6 @@ fun BaseBottomSheet(
                         .background(baseBottomSheetParams.bottomSheetContentBackgroundColor)
                         .padding(bottom = baseBottomSheetParams.bottomSheetBottomPadding)
                 ) {
-
                     if (hasCloseIcon) {
                         Image(
                             modifier = Modifier
@@ -85,7 +98,6 @@ fun BaseBottomSheet(
                             contentDescription = "Base Bottom Sheet close icon"
                         )
                     }
-
                     bottomSheetContent()
                 }
 
@@ -110,7 +122,7 @@ fun BaseBottomSheet(
         ) {
             screenContent()
 
-            backgroundBoxVisibility = sheetState.isExpanding()
+            backgroundBoxVisibility = sheetState.isExpanding() && isBackgroundDimmingEnabled
 
             AnimatedVisibility(
                 visible = backgroundBoxVisibility,
@@ -120,39 +132,13 @@ fun BaseBottomSheet(
                 Box(
                     modifier = Modifier
                         .fillMaxSize()
-                        .background(baseBottomSheetParams.backgroundColor.copy(alpha = 0.5f))
+                        .background(baseBottomSheetParams.backgroundDimmingColor.copy(alpha = 0.5f))
                         .clickable(
                             indication = null,
-                            interactionSource = remember { MutableInteractionSource() }) {
+                            interactionSource = rememberSaveable { MutableInteractionSource() }) {
                             scope.launch { sheetState.bottomSheetState.hide() }
                         })
             }
         }
     }
-}
-
-@Stable
-data class BaseBottomSheetParams(
-    val bottomSheetContentBackgroundColor: Color,
-    val topCornerRadius: Dp,
-    val bottomCornerRadius: Dp,
-    val bottomSheetBottomPadding: Dp,
-    val backgroundColor: Color,
-    val bottomSheetShadowElevation: Dp
-) {
-
-    companion object {
-
-        val Default
-            @Composable
-            get() = BaseBottomSheetParams(
-                bottomSheetContentBackgroundColor = ChiliTheme.Colors.ChiliBottomSheetBackgroundColor,
-                topCornerRadius = ChiliTheme.Attribute.ChiliBottomSheetTopCornerRadius,
-                bottomCornerRadius = ChiliTheme.Attribute.ChiliBottomSheetBottomCornerRadius,
-                backgroundColor = Color.Black,
-                bottomSheetBottomPadding = ChiliTheme.Attribute.ChiliBottomSheetContainerBottomMargin,
-                bottomSheetShadowElevation = dimensionResource(id = R.dimen.elevation_8dp)
-            )
-    }
-
 }
